+++ packages/react/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @flowkit-io/react
 
-## Unreleased
+## 1.1.0 — 2026-09-03
 
 ### Changed